+++ aiotk/__init__.py
@@ -2,10 +2,12 @@
 
 
 from .mempipe import mempipe
+from .monkey import monkey_patch
 from .posix import UnixSocketServer
 
 
 __all__ = [
     'mempipe',
+    'monkey_patch',
     'UnixSocketServer',
 ]