+++ docs/changelog.rst
@@ -6,3 +6,4 @@
 
 * :feature:`-` Adds :py:func:`aiotk.mempipe`.
 * :feature:`-` Adds :py:class:`aiotk.UnixSocketServer`.
+* :feature:`-` Adds :py:func:`aiotk.monkey_patch`.